# Faster boolean indexing and indexing by index list