--- a/docs/reference/free-functions.rst
+++ b/docs/reference/free-functions.rst
@@ -12,7 +12,6 @@ General
    bin
    bins
    bins_like
-   choose
    collapse
    group
    hist