--- a/lib/dataset/dataset_operations_common.h
+++ b/lib/dataset/dataset_operations_common.h
@@ -148,9 +148,6 @@ template <class T, class Func> DataArray transform(const T &a, Func func) {
                                                  const Dim dim);
 [[nodiscard]] Dataset strip_if_broadcast_along(const Dataset &d, const Dim dim);
 
-[[nodiscard]] DataArray strip_edges_along(const DataArray &da, const Dim dim);
-[[nodiscard]] Dataset strip_edges_along(const Dataset &ds, const Dim dim);
-
 // Helpers for reductions for DataArray and Dataset, which include masks.
 [[nodiscard]] Variable mean(const Variable &var, const Dim dim,
                             const Masks &masks);