… lib/dataset/CMakeLists.txt
… lib/dataset/include/scipp/dataset/choose.h
… lib/dataset/test/CMakeLists.txt
… lib/dataset/test/choose_test.cpp
… lib/python/CMakeLists.txt
… lib/python/choose.cpp
… lib/python/scipp.cpp
… src/scipp/core/operations.py