- Fixes #2767.
- Not equal to NumPy performance, but much better.
- Remove `sc.choose`, was was anyway in an unusable state.

#2759 already brought a big improvement for boolean indexing. In this branch we have (for dimension length of 10 million):

- Bad-case boolean: 300 ms (previously: 550 ms, NumPy: 80 ms).
- Nice-case boolean: 65 ms (NumPy: 50 ms), not much improvement
- Bad case index list: 2 s (previously: 16 s, NumPy: 0.2 s)

NumPy is still much better (especially given that it is single-threaded), but keeping the code generic and simple this may be as much as we can achieve. We can revisit this once someone has actual performance issues with this in practice.